NEX-33: Create MainMenu stub
