+++ next/components/header.tsx
@@ -1,21 +1,12 @@
 import Link from "next/link";
 import { useRouter } from "next/router";
-import { DrupalMenuLinkContent } from "next-drupal";
 import { useTranslation } from "next-i18next";
-import clsx from "clsx";
 
 import { LanguageSwitcher } from "@/components/language-switcher";
-import HamburgerIcon from "@/styles/icons/hamburger.svg";
+import { MainMenu } from "@/components/main-menu";
 import MagnifierIcon from "@/styles/icons/magnifier.svg";
 import WunderIcon from "@/styles/icons/wunder.svg";
-
-// We have applied a patch on the Drupal side that adds the langcode
-// property to the response of jsonapi menus, so we extend the type here:
-export interface DrupalMenuLinkContentWithLangcode
-  extends DrupalMenuLinkContent {
-  langcode?: string;
-  items?: DrupalMenuLinkContentWithLangcode[];
-}
+import { DrupalMenuLinkContentWithLangcode } from "@/types";
 
 interface HeaderProps {
   links: DrupalMenuLinkContentWithLangcode[];
@@ -29,7 +20,7 @@ export function Header({ links }: HeaderProps) {
         <div className="flex flex-row items-center justify-end gap-8">
           <LanguageSwitcher />
           <SearchLink />
-          <Menu items={links} />
+          <MainMenu items={links} />
         </div>
       </nav>
     </header>
@@ -55,46 +46,5 @@ function SearchLink() {
       <span className="hidden sm:mr-2 sm:inline">{t("search")}</span>
       <MagnifierIcon className="inline-block h-6 w-6" />
     </Link>
-  );
-}
-
-function Menu({ items }: { items: DrupalMenuLinkContentWithLangcode[] }) {
-  // todo: menu functionality
-  return (
-    <button onClick={() => alert("Toggle menu")}>
-      <HamburgerIcon className="inline h-6 w-6" />
-    </button>
-  );
-
-  // Only show the menu items that match the current locale:
-  const { locale } = useRouter();
-  const filteredItems = items.filter((link) => link.langcode == locale);
-  return (
-    <ul className="mx-auto mt-6 grid auto-cols-auto grid-flow-col gap-4 md:mt-0 md:auto-rows-auto md:gap-8 lg:gap-12">
-      {filteredItems.map((item) => (
-        <MenuLink link={item} key={item.id} />
-      ))}
-    </ul>
-  );
-}
-
-function MenuLink({ link }: { link: DrupalMenuLinkContentWithLangcode }) {
-  const { asPath, locale } = useRouter();
-  const actualPath = `/${locale}${asPath}`;
-
-  return (
-    <li>
-      <Link
-        href={link.url}
-        locale={locale}
-        className={clsx(
-          "py-4 text-sm hover:underline md:text-base",
-          link.url === actualPath ? "font-semibold" : "font-normal"
-        )}
-      >
-        {link.title}
-      </Link>
-      {link.items ? <Menu items={link.items} /> : null}
-    </li>
   );
 }
